Moved all code to the Internal, and just using CPS module for exporting the safe api
